requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.2
mysqlclient==2.1.1
pycodestyle==2.10.0
python-dotenv==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
